clean: use only one 'Tour' component in the dashboard
We do not want two different components UnassignedTour and Tour

recupération ficheir delete

wip remove Unassignedtourcompoone
