+++ js/app/dashboard/components/Task.js
@@ -94,11 +94,6 @@ const TaskIconRight = ({ task, onRemove }) => {
   if (task.isAssigned) {
     switch (task.status) {
     case 'TODO':
-
-      if (task.tour) {
-        return null
-      }
-
       return (
         <a
           href="#"
